Extract reusable toolbar and status components
- Move path editing UI into `EditablePathBar`
- Add shared breadcrumb, drive card, job monitor, and logo components
- Update toolbar and formatting helpers to use the new pieces
