+++ frontend/src/lib/components/settings/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Settings components barrel export
+ */
+export { default as SettingsSection } from './SettingsSection.svelte';
+export { default as SettingsRow } from './SettingsRow.svelte';
