+++ frontend/src/lib/components/ui/index.ts
@@ -8,6 +8,7 @@ export { default as ProgressButton } from './ProgressButton.svelte';
 export { default as Input } from './Input.svelte';
 export { default as Select } from './Select.svelte';
 export { default as Toggle } from './Toggle.svelte';
+export { default as Card } from './Card.svelte';
 export { default as Modal } from './Modal.svelte';
 export { default as Spinner } from './Spinner.svelte';
 export { default as Badge } from './Badge.svelte';
